+++ src/anime_parsers_ru/__init__.py
@@ -3,3 +3,4 @@ from .parser_kodik_async import KodikParserAsync
 from .parser_aniboom import AniboomParser
 from .parser_jutsu import JutsuParser
 from .parser_shikimori import ShikimoriParser
+from .parser_shikimori_async import ShikimoriParserAsync